// frontend/wallet/packages/api-layer2ledger/src/generated/models/getTransactionsResponseTransaction.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { GetTransactionsResponseTransactionLayer1TransactionId } from "./getTransactionsResponseTransactionLayer1TransactionId";
import type { GetTransactionsResponseTransactionLayer2WithdrawalId } from "./getTransactionsResponseTransactionLayer2WithdrawalId";
import type { GetTransactionsResponseTransactionSignatureDate } from "./getTransactionsResponseTransactionSignatureDate";

export interface GetTransactionsResponseTransaction {
  amount: number;
  destination_address_pubkey: string;
  fee: number;
  layer1_transaction_id?: GetTransactionsResponseTransactionLayer1TransactionId;
  layer2_withdrawal_id?: GetTransactionsResponseTransactionLayer2WithdrawalId;
  signature: string;
  signature_date?: GetTransactionsResponseTransactionSignatureDate;
  source_address_pubkey: string;
  timestamp: string;
  layer2_transaction_id: string;
  transaction_type: number;
}